+++ fields/file-list.ts
@@ -1,5 +1,6 @@
-import type { Data } from "../types.ts";
+import { transform } from "./utils.ts";
 import type {
+  Data,
   FieldDefinition,
   GroupField,
   ResolvedGroupField,
@@ -39,9 +40,7 @@ export default {
       ),
     );
 
-    // const fn = field.transform;
-    // data[field.name] = fn ? fn(value, field) : value;
-    data[field.name] = value;
+    data[field.name] = transform(field, value);
   },
 } as FieldDefinition<ResolvedFileListField>;