+++ fields/list.ts
@@ -1,4 +1,5 @@
-import { applyTextChanges } from "./utils.ts";
+import { transform } from "./utils.ts";
+import { isEmpty } from "../core/utils/string.ts";
 import type {
   FieldDefinition,
   Option,
@@ -21,7 +22,12 @@ interface ResolvedListField extends ListField, ResolvedField {
 export default {
   tag: "f-list",
   jsImport: "lume_cms/components/f-list.js",
-  applyChanges: applyTextChanges,
+  applyChanges(data, changes, field) {
+    const value = Object.values(changes[field.name] || {}).filter((v) =>
+      !isEmpty(v)
+    );
+    data[field.name] = transform(field, value);
+  },
 } as FieldDefinition<ResolvedListField>;
 
 declare global {